feat: add lambdas build and zip
