+++ aws/lambdas/process-pending-notifications.cjs
@@ -1,0 +1,1 @@
+"use strict";var y=Object.defineProperty;var W=Object.getOwnPropertyDescriptor;var K=Object.getOwnPropertyNames;var M=Object.prototype.hasOwnProperty;var A=(o,r)=>{for(var _ in r)y(o,_,{get:r[_],enumerable:!0})},Q=(o,r,_,p)=>{if(r&&typeof r=="object"||typeof r=="function")for(let N of K(r))!M.call(o,N)&&N!==_&&y(o,N,{get:()=>r[N],enumerable:!(p=W(r,N))||p.enumerable});return o};var B=o=>Q(y({},"__esModule",{value:!0}),o);var G={};A(G,{handler:()=>q});module.exports=B(G);var j=require("dotenv/config"),t=require("zod"),L=t.z.object({API_PORT:t.z.coerce.number(),NODE_ENV:t.z.enum(["development","production"]),JWT_SECRET:t.z.string(),JWT_EXPIRATION:t.z.coerce.number(),POSTGRES_USER:t.z.string(),POSTGRES_PASSWORD:t.z.string(),POSTGRES_DB:t.z.string(),DATABASE_URL:t.z.string(),DRIZZLE_KIT_DATABASE_URL:t.z.string(),PASSWORD_SALT_ROUNDS:t.z.coerce.number(),AWS_SQS_QUEUE_URL:t.z.string().url(),AWS_SQS_REGION:t.z.string(),AWS_SQS_ACCESS_KEY_ID:t.z.string(),AWS_SQS_SECRET_ACCESS_KEY:t.z.string(),AWS_SNS_ACCESS_KEY_ID:t.z.string(),AWS_SNS_SECRET_ACCESS_KEY:t.z.string(),AWS_SNS_REGION:t.z.string(),RESEND_API_KEY:t.z.string(),SMS_VERIFICATION_CODE_EXPIRES_IN:t.z.coerce.number(),EMAIL_VERIFICATION_CODE_EXPIRES_IN:t.z.coerce.number()}),f=Object.freeze(L.parse(process.env));var P=require("date-fns"),u=require("drizzle-orm");var D=require("drizzle-orm/node-postgres");var E={};A(E,{notificationRelations:()=>k,notificationStatusEnum:()=>w,notifications:()=>e,products:()=>l,productsRelations:()=>F,users:()=>c,usersRelations:()=>V,validationTypesEnum:()=>C,validations:()=>Y});var h=require("drizzle-orm"),i=require("drizzle-orm/pg-core"),x=require("crypto");var T=require("drizzle-orm"),n=require("drizzle-orm/pg-core"),R=require("crypto");var s=require("drizzle-orm/pg-core"),I=require("crypto"),g=require("drizzle-orm");var w=(0,s.pgEnum)("notification_status",["created","scheduled","sent","failed"]),e=(0,s.pgTable)("notifications",{id:(0,s.text)("id").$defaultFn(I.randomUUID).primaryKey(),userId:(0,s.text)("user_id").notNull().references(()=>c.id),productId:(0,s.text)("product_id").notNull().references(()=>l.id,{onDelete:"cascade"}),notifyAt:(0,s.timestamp)("notify_at",{withTimezone:!0,mode:"string"}).notNull(),status:w("status").notNull().default("created"),createdAt:(0,s.timestamp)("created_at",{withTimezone:!0}).defaultNow(),updatedAt:(0,s.timestamp)("updated_at",{withTimezone:!0}).defaultNow()}),k=(0,g.relations)(e,({one:o})=>({product:o(l,{fields:[e.productId],references:[l.id]}),user:o(c,{fields:[e.userId],references:[c.id]})}));var l=(0,n.pgTable)("products",{id:(0,n.text)("id").$defaultFn(R.randomUUID).primaryKey(),userId:(0,n.text)("user_id").notNull().references(()=>c.id),name:(0,n.text)("name").notNull(),buyedAt:(0,n.date)("buyed_at",{mode:"string"}).notNull(),expiresAt:(0,n.date)("expires_at",{mode:"string"}).notNull(),category:(0,n.text)("category").notNull(),updatedAt:(0,n.timestamp)("updated_at",{withTimezone:!0,mode:"date"}).notNull().defaultNow(),createdAt:(0,n.timestamp)("created_at",{withTimezone:!0,mode:"date"}).notNull().defaultNow()}),F=(0,T.relations)(l,({one:o,many:r})=>({user:o(c,{fields:[l.userId],references:[c.id]}),notifications:r(e)}));var c=(0,i.pgTable)("users",{id:(0,i.text)("id").$defaultFn(x.randomUUID).primaryKey(),isPhoneVerified:(0,i.boolean)("is_phone_verified").notNull().default(!1),isEmailVerified:(0,i.boolean)("is_email_verified").notNull().default(!1),name:(0,i.text)("name").notNull(),email:(0,i.text)("email").notNull().unique(),phoneNumber:(0,i.text)("phone_number").notNull().unique(),password:(0,i.text)("password").notNull(),avatarUrl:(0,i.text)("avatar_url"),updatedAt:(0,i.timestamp)("updated_at",{withTimezone:!0,mode:"date"}).notNull().defaultNow(),createdAt:(0,i.timestamp)("created_at",{withTimezone:!0,mode:"date"}).notNull().defaultNow(),deletedAt:(0,i.timestamp)("deleted_at",{withTimezone:!0,mode:"date"})}),V=(0,h.relations)(c,({many:o})=>({products:o(l)}));var d=require("drizzle-orm/pg-core"),C=(0,d.pgEnum)("validation_types",["phone","email"]),Y=(0,d.pgTable)("validations",{code:(0,d.text)("code").notNull().primaryKey(),type:C("type").notNull(),identifier:(0,d.text)("identifier").notNull().unique(),usedAt:(0,d.timestamp)("used_at",{withTimezone:!0,mode:"date"}),expiresAt:(0,d.timestamp)("expires_at",{withTimezone:!0,mode:"date"}).notNull(),createdAt:(0,d.timestamp)("created_at",{withTimezone:!0,mode:"date"}).notNull().defaultNow()});var S=(0,D.drizzle)(f.DATABASE_URL,{schema:E,logger:f.NODE_ENV==="development"});var U=()=>({createNotification:async a=>{let[m]=await S.insert(e).values(a).returning();return m},getNotificationById:async a=>{let[m]=await S.select().from(e).where((0,u.eq)(e.id,a));return m},updateNotificationById:async(a,m)=>{let[z]=await S.update(e).set({...m}).where((0,u.eq)(e.id,a)).returning();return z},getPendingNotifications:async()=>{let m=(0,P.format)(new Date,"yyyy-MM-dd");return await S.select().from(e).where((0,u.and)((0,u.eq)(e.status,"created"),(0,u.eq)(e.notifyAt,m))).execute()},scheduleNotification:async a=>{let[m]=await S.update(e).set({status:"scheduled"}).where((0,u.eq)(e.id,a)).returning();return m},markNotificationAsSent:async a=>await S.update(e).set({status:"sent"}).where((0,u.eq)(e.id,a)),markNotificationAsFailed:async a=>await S.update(e).set({status:"failed"}).where((0,u.eq)(e.id,a))});var O=require("@aws-sdk/client-sqs");var b=new O.SQSClient({region:f.AWS_SQS_REGION,credentials:{accessKeyId:f.AWS_SQS_ACCESS_KEY_ID,secretAccessKey:f.AWS_SQS_SECRET_ACCESS_KEY}});var v=require("@aws-sdk/client-sqs"),q=async()=>{try{let{getPendingNotifications:o,scheduleNotification:r}=U(),_=await o();for(let p of _)await b.send(new v.SendMessageCommand({QueueUrl:f.AWS_SQS_QUEUE_URL,MessageBody:JSON.stringify({notificationId:p.id,userId:p.userId,productId:p.productId})})),await r(p.id);console.log("Processed pending notifications")}catch(o){console.error("Error processing pending notifications:",o)}};0&&(module.exports={handler});
