+++ src/modules/auth/routes.ts
@@ -4,14 +4,14 @@ import { authenticateUserRequestSchema } from './authenticate/schemas.ts';
 import { authenticateUser } from './authenticate/use-case.ts';
 
 import { validateSchema } from '@/shared/helpers/validate-schema.ts';
-import { sendEmailVerificationCodeRequestSchema } from './verification/send-email-verification-code/schemas.ts';
-import { sendEmailVerificationCode } from './verification/send-email-verification-code/use-case.ts';
-import { sendPhoneVerificationCodeRequestSchema } from './verification/send-phone-verification-code/schemas.ts';
-import { sendPhoneVerificationCode } from './verification/send-phone-verification-code/use-case.ts';
-import { verifyEmailRequestSchema } from './verification/verify-email/schemas.ts';
-import { verifyEmail } from './verification/verify-email/use-case.ts';
-import { verifyPhoneNumberRequestSchema } from './verification/verify-phone-number/schemas.ts';
-import { verifyPhoneNumber } from './verification/verify-phone-number/use-case.ts';
+import { sendEmailVerificationCodeRequestSchema } from './send-email-verification-code/schemas.ts';
+import { sendEmailVerificationCode } from './send-email-verification-code/use-case.ts';
+import { sendPhoneVerificationCodeRequestSchema } from './send-phone-verification-code/schemas.ts';
+import { sendPhoneVerificationCode } from './send-phone-verification-code/use-case.ts';
+import { verifyEmailRequestSchema } from './verify-email/schemas.ts';
+import { verifyEmail } from './verify-email/use-case.ts';
+import { verifyPhoneNumberRequestSchema } from './verify-phone-number/schemas.ts';
+import { verifyPhoneNumber } from './verify-phone-number/use-case.ts';
 export const authRoutes = new Hono();
 
 authRoutes.post('/login', async (c) => {
