+++ src/modules/auth/send-email-verification-code/schemas.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const sendEmailVerificationCodeRequestSchema = z.object({
+  name: z.string().min(1),
+  email: z.string().email()
+});
+
+export type SendEmailVerificationCodeInput = z.infer<typeof sendEmailVerificationCodeRequestSchema>;
