+++ src/modules/auth/send-phone-verification-code/schemas.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+import { phoneNumberRegex } from '@/shared/constants/phone.ts';
+
+export const sendPhoneVerificationCodeRequestSchema = z.object({
+  phoneNumber: z.string().regex(phoneNumberRegex, 'Invalid phone number')
+});
+
+export type SendPhoneVerificationCodeInput = z.infer<typeof sendPhoneVerificationCodeRequestSchema>;
