--- src/modules/auth/verification/verify-phone-number/schemas.ts
@@ -1,9 +1,0 @@
-import { phoneNumberRegex } from '@/shared/constants/phone.ts';
-import { z } from 'zod';
-
-export const verifyPhoneNumberRequestSchema = z.object({
-  phoneNumber: z.string().regex(phoneNumberRegex, 'Invalid phone number'),
-  verificationCode: z.string().length(6)
-});
-
-export type VerifyPhoneNumberInput = z.infer<typeof verifyPhoneNumberRequestSchema>;
